# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/audit": {
      "get": {
        "description": "Возвращает записи аудита с возможностью фильтрации по таблице, операции, пользователю и ID записи",
        "operationId": "AuditController_getAuditLogs",
        "parameters": [
          {
            "name": "table",
            "required": false,
            "in": "query",
            "description": "Имя таблицы для фильтрации",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "operation",
            "required": false,
            "in": "query",
            "description": "Тип операции (INSERT, UPDATE, DELETE)",
            "schema": {
              "enum": [
                "INSERT",
                "UPDATE",
                "DELETE"
              ],
              "type": "string"
            }
          },
          {
            "name": "user",
            "required": false,
            "in": "query",
            "description": "Пользователь, выполнивший операцию",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "required": false,
            "in": "query",
            "description": "ID измененной записи",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Лимит записей (по умолчанию 50)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Записи аудита успешно получены"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить записи аудита с фильтрацией",
        "tags": [
          "audit"
        ]
      }
    },
    "/audit/recent": {
      "get": {
        "description": "Возвращает последние записи аудита в порядке убывания по времени",
        "operationId": "AuditController_getRecentAuditLogs",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Количество записей (по умолчанию 50)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Последние записи аудита получены"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить последние записи аудита",
        "tags": [
          "audit"
        ]
      }
    },
    "/audit/history": {
      "get": {
        "operationId": "AuditController_getRecordHistory",
        "parameters": [
          {
            "name": "table",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
    "/audit/statistics": {
      "get": {
        "description": "Возвращает статистику операций аудита по таблицам и типам операций",
        "operationId": "AuditController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Статистика аудита получена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "operationStats": {
                      "type": "array",
                      "description": "Статистика по операциям"
                    },
                    "tableStats": {
                      "type": "array",
                      "description": "Статистика по таблицам"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить статистику аудита",
        "tags": [
          "audit"
        ]
      }
    },
    "/audit/recent-changes": {
      "get": {
        "operationId": "AuditController_getRecentChanges",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
    "/audit/user-activity": {
      "get": {
        "operationId": "AuditController_getUserActivity",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials"
          }
        },
        "summary": "User login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation failed"
          },
          "409": {
            "description": "Conflict - Username or email already exists"
          }
        },
        "summary": "User registration",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User logout",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Georgian Restaurant API",
    "description": "API для управления грузинскими ресторанами с системой бронирования, заказов и аудита",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username or email",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "type": "object",
            "description": "User information",
            "example": {
              "user_id": "uuid-here",
              "username": "admin",
              "email": "admin@georgian-restaurant.ge",
              "first_name": "Тенгиз",
              "last_name": "Админидзе",
              "role": "admin",
              "role_id": 1
            }
          }
        },
        "required": [
          "access_token",
          "user"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}